renamed api_gateway.py and updated imports
