+++ test/unit_tests/test_api_gateway.py
@@ -1,6 +1,6 @@
 from amazonia.classes.api_gateway_config import ApiGatewayResponseConfig, ApiGatewayRequestConfig
 from amazonia.classes.api_gateway_config import ApiGatewayMethodConfig
-from amazonia.classes.api_gateway import ApiGatewayUnit
+from amazonia.classes.api_gateway_unit import ApiGatewayUnit
 from troposphere import Template, Ref, Join
 from nose.tools import *
